Bai tap 10: Protocol bai 3 + 4
