+++ 10_Protocol/PhuocN/Baitap3+4/PracticeTemplate/Protocol/Models/Objects/Huyen.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class Huyen: Location {
+class Huyen {
     var name: String
     
     init(name: String) {
