+++ 10_Protocol/PhuocN/Baitap3+4/PracticeTemplate/Protocol/Models/Objects/Mien.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class Mien: Location {
+class Mien {
     var name: String
     var tinh: [Tinh]
     
@@ -17,3 +17,129 @@ class Mien: Location {
         self.tinh = tinh
     }
 }
+
+let miens: [Mien] = [Mien(name: "Mien 1", tinh: [Tinh(name: "Tinh 1.1", huyen: [Huyen(name: "Huyen 1.1.1"),
+                                                                                Huyen(name: "Huyen 1.1.2"),
+                                                                                Huyen(name: "Huyen 1.1.3"),
+                                                                                Huyen(name: "Huyen 1.1.4"),
+                                                                                Huyen(name: "Huyen 1.1.5")]),
+                                                 Tinh(name: "Tinh 1.2", huyen: [Huyen(name: "Huyen 1"),
+                                                                                Huyen(name: "Huyen 1.2.1"),
+                                                                                Huyen(name: "Huyen 1.2.3"),
+                                                                                Huyen(name: "Huyen 1.2.4"),
+                                                                                Huyen(name: "Huyen 1.2.5")]),
+                                                 Tinh(name: "Tinh 1.3", huyen: [Huyen(name: "Huyen 1.3.1"),
+                                                                                Huyen(name: "Huyen 1.3.2"),
+                                                                                Huyen(name: "Huyen 1.3.3"),
+                                                                                Huyen(name: "Huyen 1.3.4"),
+                                                                                Huyen(name: "Huyen 1.3.5")]),
+                                                 Tinh(name: "Tinh 1.4", huyen: [Huyen(name: "Huyen 1.4.1"),
+                                                                                Huyen(name: "Huyen 1.4.2"),
+                                                                                Huyen(name: "Huyen 1.4.3"),
+                                                                                Huyen(name: "Huyen 1.4.4"),
+                                                                                Huyen(name: "Huyen 1.4.5")]),
+                                                 Tinh(name: "Tinh 1.5", huyen: [Huyen(name: "Huyen 1.5.1"),
+                                                                                Huyen(name: "Huyen 1.5.2"),
+                                                                                Huyen(name: "Huyen 1.5.3"),
+                                                                                Huyen(name: "Huyen 1.5.4"),
+                                                                                Huyen(name: "Huyen 1.5.5")])]),
+                     Mien(name: "Mien 2", tinh: [Tinh(name: "Tinh 2.1", huyen: [Huyen(name: "Huyen 2.1.1"),
+                                                                                Huyen(name: "Huyen 2.1.2"),
+                                                                                Huyen(name: "Huyen 2.1.3"),
+                                                                                Huyen(name: "Huyen 2.1.4"),
+                                                                                Huyen(name: "Huyen 2.1.5")]),
+                                                 Tinh(name: "Tinh 2.2", huyen: [Huyen(name: "Huyen 2.2.1"),
+                                                                                Huyen(name: "Huyen 2.2.2"),
+                                                                                Huyen(name: "Huyen 2.2.3"),
+                                                                                Huyen(name: "Huyen 2.2.4"),
+                                                                                Huyen(name: "Huyen 2.2.5")]),
+                                                 Tinh(name: "Tinh 2.3", huyen: [Huyen(name: "Huyen 2.3.1"),
+                                                                                Huyen(name: "Huyen 2.3.2"),
+                                                                                Huyen(name: "Huyen 2.3.3"),
+                                                                                Huyen(name: "Huyen 2.3.4"),
+                                                                                Huyen(name: "Huyen 2.3.5")]),
+                                                 Tinh(name: "Tinh 2.4", huyen: [Huyen(name: "Huyen 2.4.1"),
+                                                                                Huyen(name: "Huyen 2.4.2"),
+                                                                                Huyen(name: "Huyen 2.4.3"),
+                                                                                Huyen(name: "Huyen 2.4.4"),
+                                                                                Huyen(name: "Huyen 2.4.5")]),
+                                                 Tinh(name: "Tinh 2.5", huyen: [Huyen(name: "Huyen 2.5.1"),
+                                                                                Huyen(name: "Huyen 2.5.2"),
+                                                                                Huyen(name: "Huyen 2.5.3"),
+                                                                                Huyen(name: "Huyen 2.5.4"),
+                                                                                Huyen(name: "Huyen 2.5.5")])]),
+                     Mien(name: "Mien 3", tinh: [Tinh(name: "Tinh 3.1", huyen: [Huyen(name: "Huyen 3.1.1"),
+                                                                                Huyen(name: "Huyen 3.1.2"),
+                                                                                Huyen(name: "Huyen 3.1.3"),
+                                                                                Huyen(name: "Huyen 3.1.4"),
+                                                                                Huyen(name: "Huyen 3.1.5")]),
+                                                 Tinh(name: "Tinh 3.2", huyen: [Huyen(name: "Huyen 3.2.1"),
+                                                                                Huyen(name: "Huyen 3.2.2"),
+                                                                                Huyen(name: "Huyen 3.2.3"),
+                                                                                Huyen(name: "Huyen 3.2.4"),
+                                                                                Huyen(name: "Huyen 3.2.5")]),
+                                                 Tinh(name: "Tinh 3.3", huyen: [Huyen(name: "Huyen 3.3.1"),
+                                                                                Huyen(name: "Huyen 3.3.2"),
+                                                                                Huyen(name: "Huyen 3.3.3"),
+                                                                                Huyen(name: "Huyen 3.3.4"),
+                                                                                Huyen(name: "Huyen 3.3.5")]),
+                                                 Tinh(name: "Tinh 3.4", huyen: [Huyen(name: "Huyen 3.4.1"),
+                                                                                Huyen(name: "Huyen 3.4.2"),
+                                                                                Huyen(name: "Huyen 3.4.3"),
+                                                                                Huyen(name: "Huyen 3.4.4"),
+                                                                                Huyen(name: "Huyen 3.4.5")]),
+                                                 Tinh(name: "Tinh 3.5", huyen: [Huyen(name: "Huyen 3.5.1"),
+                                                                                Huyen(name: "Huyen 3.5.2"),
+                                                                                Huyen(name: "Huyen 3.5.3"),
+                                                                                Huyen(name: "Huyen 3.5.4"),
+                                                                                Huyen(name: "Huyen 3.5.5")])]),
+                     Mien(name: "Mien 4", tinh: [Tinh(name: "Tinh 4.1", huyen: [Huyen(name: "Huyen 4.1.1"),
+                                                                                Huyen(name: "Huyen 4.1.2"),
+                                                                                Huyen(name: "Huyen 4.1.3"),
+                                                                                Huyen(name: "Huyen 4.1.4"),
+                                                                                Huyen(name: "Huyen 4.1.5")]),
+                                                 Tinh(name: "Tinh 4.2", huyen: [Huyen(name: "Huyen 4.2.1"),
+                                                                                Huyen(name: "Huyen 4.2.2"),
+                                                                                Huyen(name: "Huyen 4.2.3"),
+                                                                                Huyen(name: "Huyen 4.2.4"),
+                                                                                Huyen(name: "Huyen 4.2.5")]),
+                                                 Tinh(name: "Tinh 4.3", huyen: [Huyen(name: "Huyen 4.3.1"),
+                                                                                Huyen(name: "Huyen 4.3.2"),
+                                                                                Huyen(name: "Huyen 4.3.3"),
+                                                                                Huyen(name: "Huyen 4.3.4"),
+                                                                                Huyen(name: "Huyen 4.3.5")]),
+                                                 Tinh(name: "Tinh 4.4", huyen: [Huyen(name: "Huyen 4.4.1"),
+                                                                                Huyen(name: "Huyen 4.4.2"),
+                                                                                Huyen(name: "Huyen 4.4.3"),
+                                                                                Huyen(name: "Huyen 4.4.4"),
+                                                                                Huyen(name: "Huyen 4.4.5")]),
+                                                 Tinh(name: "Tinh 4.5", huyen: [Huyen(name: "Huyen 4.5.1"),
+                                                                                Huyen(name: "Huyen 4.5.2"),
+                                                                                Huyen(name: "Huyen 4.5.3"),
+                                                                                Huyen(name: "Huyen 4.5.4"),
+                                                                                Huyen(name: "Huyen 4.5.5")])]),
+                     Mien(name: "Mien 5", tinh: [Tinh(name: "Tinh 5.1", huyen: [Huyen(name: "Huyen 5.1.1"),
+                                                                                Huyen(name: "Huyen 5.1.2"),
+                                                                                Huyen(name: "Huyen 5.1.3"),
+                                                                                Huyen(name: "Huyen 5.1.4"),
+                                                                                Huyen(name: "Huyen 5.1.5")]),
+                                                 Tinh(name: "Tinh 5.2", huyen: [Huyen(name: "Huyen 5.2.1"),
+                                                                                Huyen(name: "Huyen 5.2.2"),
+                                                                                Huyen(name: "Huyen 5.2.3"),
+                                                                                Huyen(name: "Huyen 5.2.4"),
+                                                                                Huyen(name: "Huyen 5.2.5")]),
+                                                 Tinh(name: "Tinh 5.3", huyen: [Huyen(name: "Huyen 5.3.1"),
+                                                                                Huyen(name: "Huyen 5.3.2"),
+                                                                                Huyen(name: "Huyen 5.3.3"),
+                                                                                Huyen(name: "Huyen 5.3.4"),
+                                                                                Huyen(name: "Huyen 5.3.5")]),
+                                                 Tinh(name: "Tinh 5.4", huyen: [Huyen(name: "Huyen 5.4.1"),
+                                                                                Huyen(name: "Huyen 5.4.2"),
+                                                                                Huyen(name: "Huyen 5.4.3"),
+                                                                                Huyen(name: "Huyen 5.4.4"),
+                                                                                Huyen(name: "Huyen 5.4.5")]),
+                                                 Tinh(name: "Tinh 5.5", huyen: [Huyen(name: "Huyen 5.5.1"),
+                                                                                Huyen(name: "Huyen 5.5.2"),
+                                                                                Huyen(name: "Huyen 5.5.3"),
+                                                                                Huyen(name: "Huyen 5.5.4"),
+                                                                                Huyen(name: "Huyen 5.5.5")])])]
