+++ 10_Protocol/PhuocN/Baitap3+4/PracticeTemplate/Protocol/Models/Objects/Tinh.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class Tinh: Location {
+class Tinh {
     var name: String
     var huyen: [Huyen]
